+++ packages/@expo/cli/CHANGELOG.md
@@ -16,6 +16,8 @@
 
 ### 💡 Others
 
+- [Internal] Remove RSC render context ([#45908](https://github.com/expo/expo/pull/45908) by [@kitten](https://github.com/kitten))
+
 ## 55.0.30 — 2026-05-13
 
 _This version does not introduce any user-facing changes._