+++ packages/@expo/router-server/CHANGELOG.md
@@ -13,6 +13,8 @@
 
 ### 💡 Others
 
+- [Internal] Remove RSC render context and remove render store global ([#45908](https://github.com/expo/expo/pull/45908) by [@kitten](https://github.com/kitten))
+
 ## 55.0.16 — 2026-05-05
 
 _This version does not introduce any user-facing changes._